Moving `scratchpad_xyhw` to models
* Moving `Scratchpad` to models

* Moving `scratchpad_xyhw` to models

* Renaming `scratchpad_xyhw` to `Scratchpad::xyhw`
