+++ leftwm-core/src/models/mod.rs
@@ -6,6 +6,7 @@ mod layout_manager;
 mod manager;
 mod margins;
 mod mode;
+mod scratchpad;
 mod screen;
 mod size;
 mod tag;
@@ -30,6 +31,7 @@ pub use layout_manager::LayoutMode;
 pub use manager::Manager;
 pub use margins::Margins;
 pub use mode::Mode;
+pub use scratchpad::ScratchPad;
 pub use screen::{BBox, Screen};
 pub use size::Size;
 pub use window::Window;
