+++ leftwm-core/src/models/scratchpad.rs
@@ -1,0 +1,45 @@
+use crate::models::Size;
+use serde::{Deserialize, Serialize};
+
+use super::{Xyhw, XyhwBuilder};
+
+#[derive(Serialize, Deserialize, Debug, Clone, PartialEq)]
+pub struct ScratchPad {
+    pub name: String,
+    pub value: String,
+    // relative x of scratchpad, 25 means 25% of workspace x
+    pub x: Option<Size>,
+    // relative y of scratchpad, 25 means 25% of workspace y
+    pub y: Option<Size>,
+    // relative height of scratchpad, 50 means 50% of workspace height
+    pub height: Option<Size>,
+    // relative width of scratchpad, 50 means 50% of workspace width
+    pub width: Option<Size>,
+}
+
+impl ScratchPad {
+    // Get size and position of scratchpad from config and workspace size.
+    pub fn xyhw(&self, xyhw: &Xyhw) -> Xyhw {
+        let x_sane = sane_dimension(self.x, 0.25, xyhw.w());
+        let y_sane = sane_dimension(self.y, 0.25, xyhw.h());
+        let height_sane = sane_dimension(self.height, 0.50, xyhw.h());
+        let width_sane = sane_dimension(self.width, 0.50, xyhw.w());
+
+        XyhwBuilder {
+            x: xyhw.x() + x_sane,
+            y: xyhw.y() + y_sane,
+            h: height_sane,
+            w: width_sane,
+            ..XyhwBuilder::default()
+        }
+        .into()
+    }
+}
+
+fn sane_dimension(config_value: Option<Size>, default_ratio: f32, max_pixel: i32) -> i32 {
+    match config_value {
+        Some(size @ Size::Ratio(r)) if (0.0..=1.0).contains(&r) => size.into_absolute(max_pixel),
+        Some(Size::Pixel(pixel)) if (0..=max_pixel).contains(&pixel) => pixel,
+        _ => Size::Ratio(default_ratio).into_absolute(max_pixel),
+    }
+}
